move xr to playground
